backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.5
djoser==2.1.0
Pillow==9.5.0
psycopg2-binary==2.9.3